fix references to moved files
