+++ modules/engines/engine-hdb/src/test/java/com/sap/xsk/hdb/ds/test/handlers/HDBTableAlterHandlerTest.java
@@ -19,7 +19,7 @@ import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableColumnModel;
 import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableConstraintPrimaryKeyModel;
 import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableConstraintsModel;
 import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableModel;
-import com.sap.xsk.hdb.ds.processors.table.utils.XSKTableAlterHandler;
+import com.sap.xsk.hdb.ds.processors.table.XSKTableAlterHandler;
 import java.sql.Connection;
 import java.sql.DatabaseMetaData;
 import java.sql.ResultSet;
