+++ modules/engines/engine-hdb/src/test/java/com/sap/xsk/hdb/ds/test/processors/HDBTableProcessorsTest.java
@@ -16,8 +16,10 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
+import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableModel;
+import com.sap.xsk.hdb.ds.processors.table.XSKTableAlterHandler;
+import com.sap.xsk.hdb.ds.processors.table.XSKTableAlterProcessor;
 import java.sql.Connection;
-
 import org.eclipse.dirigible.core.test.AbstractDirigibleTest;
 import org.junit.Before;
 import org.junit.Test;
@@ -28,46 +30,42 @@ import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import com.sap.xsk.hdb.ds.model.hdbtable.XSKDataStructureHDBTableModel;
-import com.sap.xsk.hdb.ds.processors.table.XSKTableAlterProcessor;
-import com.sap.xsk.hdb.ds.processors.table.utils.XSKTableAlterHandler;
-
 @RunWith(MockitoJUnitRunner.class)
 public class HDBTableProcessorsTest extends AbstractDirigibleTest {
 
-	@Mock(answer = Answers.RETURNS_DEEP_STUBS)
-	private Connection mockConnection;
+  @Mock(answer = Answers.RETURNS_DEEP_STUBS)
+  private Connection mockConnection;
 
-	@Mock
-	private XSKDataStructureHDBTableModel mockModel;
+  @Mock
+  private XSKDataStructureHDBTableModel mockModel;
 
-	@Mock
-	private XSKTableAlterHandler mockHandler;
+  @Mock
+  private XSKTableAlterHandler mockHandler;
 
-	private XSKTableAlterProcessor xskTableAlterProcessor = spy(new XSKTableAlterProcessor());
+  private XSKTableAlterProcessor xskTableAlterProcessor = spy(new XSKTableAlterProcessor());
 
-	@Before
-	public void openMocks() {
-		MockitoAnnotations.openMocks(this);
-	}
+  @Before
+  public void openMocks() {
+    MockitoAnnotations.openMocks(this);
+  }
 
-	@Test
-	public void executeAlterSuccessfully() throws Exception {
-		Mockito.doReturn(mockHandler).when(xskTableAlterProcessor).createTableAlterHandler(mockConnection, mockModel);
+  @Test
+  public void executeAlterSuccessfully() throws Exception {
+    Mockito.doReturn(mockHandler).when(xskTableAlterProcessor).createTableAlterHandler(mockConnection, mockModel);
 
-		doNothing().when(mockHandler).addColumns(mockConnection);
-		doNothing().when(mockHandler).removeColumns(mockConnection);
-		doNothing().when(mockHandler).updateColumns(mockConnection);
-		doNothing().when(mockHandler).rebuildIndeces(mockConnection);
-		doNothing().when(mockHandler).ensurePrimaryKeyIsUnchanged(mockConnection);
+    doNothing().when(mockHandler).addColumns(mockConnection);
+    doNothing().when(mockHandler).removeColumns(mockConnection);
+    doNothing().when(mockHandler).updateColumns(mockConnection);
+    doNothing().when(mockHandler).rebuildIndeces(mockConnection);
+    doNothing().when(mockHandler).ensurePrimaryKeyIsUnchanged(mockConnection);
 
-		xskTableAlterProcessor.execute(mockConnection, mockModel);
+    xskTableAlterProcessor.execute(mockConnection, mockModel);
 
-		verify(mockHandler, times(1)).addColumns(mockConnection);
-		verify(mockHandler, times(1)).removeColumns(mockConnection);
-		verify(mockHandler, times(1)).updateColumns(mockConnection);
-		verify(mockHandler, times(1)).rebuildIndeces(mockConnection);
-		verify(mockHandler, times(1)).ensurePrimaryKeyIsUnchanged(mockConnection);
-	}
+    verify(mockHandler, times(1)).addColumns(mockConnection);
+    verify(mockHandler, times(1)).removeColumns(mockConnection);
+    verify(mockHandler, times(1)).updateColumns(mockConnection);
+    verify(mockHandler, times(1)).rebuildIndeces(mockConnection);
+    verify(mockHandler, times(1)).ensurePrimaryKeyIsUnchanged(mockConnection);
+  }
 
 }
